refactor to record class
