+++ src/main/java/org/broadinstitute/consent/http/models/Error.java
@@ -1,0 +1,3 @@
+package org.broadinstitute.consent.http.models;
+
+public record Error(String message, Integer code) {}
