--- src/main/java/org/broadinstitute/consent/http/models/dto/Error.java
@@ -1,32 +1,0 @@
-package org.broadinstitute.consent.http.models.dto;
-
-public class Error {
-
-    private String message;
-
-    private Integer code;
-
-    public Error() {
-    }
-
-    public Error(String message, Integer code) {
-        this.message = message;
-        this.code = code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String error) {
-        this.message = error;
-    }
-
-    public Integer getCode() {
-        return code;
-    }
-
-    public void setCode(Integer code) {
-        this.code = code;
-    }
-}
