+++ src/main/java/org/broadinstitute/consent/http/resources/ErrorResource.java
@@ -6,7 +6,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
-import org.broadinstitute.consent.http.models.dto.Error;
+import org.broadinstitute.consent.http.models.Error;
 import org.eclipse.jetty.server.Request;
 
 @Path("error")
@@ -19,6 +19,6 @@ public class ErrorResource {
     String originalUri = ((Request) request).getOriginalURI();
     String msg = String.format("Unable to find requested path: '%s'", originalUri);
     Error error = new Error(msg, 404);
-    return Response.status(error.getCode()).entity(error).build();
+    return Response.status(error.code()).entity(error).build();
   }
 }
